// v2/api-validator/src/client/generated/models/WithdrawalAddressPolicyRequirement.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type WithdrawalAddressPolicyRequirement = {
    value: WithdrawalAddressPolicyRequirement.value;
};

export namespace WithdrawalAddressPolicyRequirement {

    export enum value {
        APPROVED_ADDRESSES_ONLY = 'ApprovedAddressesOnly',
        ALL_ADDRESSES = 'AllAddresses',
    }


}